fixed scipy deprication warning, silenced numba
scipy function minres deprecated keyword 'tol' for 'rtol', those were updated.  Numba is having a conflict with the data being passed to build_sigma_reduced5; this wasn't a problem before the updgrade to mac os sequoia.  I am not sure why this is a problem now, but I am silencing it by commenting out the @jit decorator for this function.  We should try to fix it later.